Lesson 27 - Adding Recipe Docs
Added functionality to db.js to add documents to 'recipes' collection in Firestore by setting listener on form tag and performing add() method on collection when form is submitted. To capture updated changes in db.js, changed dynamicCacheName to force update on cache. Updated README with notes regarding offline use of IndexedDB with app and how to add documents to Firestore.